Move component to component package. Move componentShortFlag also. Not sure why it was on delete and not on component.
